Update changelog, split tests
